🛠️ Refactor: main과 add 기능 모듈화
